Revert "Update TetrisPlayManager.h"
This reverts commit 2f77117e914db46aa12ab572fa02eaae38bde150.

+++ Source/MultiplayerTetris/Public/TetrisPlayManager.h
@@ -26,7 +26,7 @@ enum class ELockDownOption : uint8
 };
 
 UENUM()
-enum class EPhase : uint16
+enum class EPhase : uint8
 {
 	None,
 	Generation,
